[PATCH] Automatic update of the duration of the weight raising
To provide low latency, bfq raises the weight of the applications to be privileged. This heuristic alone does not suffice. One of the other additional mechanisms used by bfq is performing disk idling for weight-raised applications. Disk idling means that, if one such applications runs out of requests while it is being served, bfq idle the disk and waits for a few ms for the next request of the application to arrive.

This waiting is critical for low latency, but does not allow SSD devices to reach maximum throughput.
For this reason, the weight-raising (and related disk-idling) duration should be kept low. For interactive applications, this duration is controlled by the raising_max_time sysfs parameter. We set its value, conservatively, to 7.5 seconds.

Once applied this experimental patch things change. If raising_max_time is set to 0 (default value), then bfq computes it automatically. In particular, bfq adapts, heuristically, to the speed of the disk: the higher the speed of the disk is, the lower it takes to load applications on that disk, and hence the lower the value assigned automatically by bfq to this parameter is. By doing so, after a new application is started, bfq goes into 'high-thoughput for SSD' mode for that application as soon as possible, without sacrificing latency. You can find more detail sin the comments within the patch itself.

It is of course an heuristic, so only usage can tell us how effective it is actually. According to our tests, it is apparently working well (we are just using it on our PCs). This patch is for v4-3.5, but it  probably applies (almost) cleanly also to v4 for older kernel versions. In case of problems just let us know.

https://groups.google.com/forum/?fromgroups#!topic/bfq-iosched/hPe2jFW55Is%5B1-25%5D